Updates and integration with aws-lambda-redshift-loader
Couple of points on this codebase:

* There's no massive benefit to doing an S3 copy over just doing a
PutObject each time, and it makes the code a lot more simple - so have
removed the copy function, the need for a headObject, etc
* You were doing a Scan on the DDB table, and then for each prefix
potentially doing a PutObject and within it's callback calling
context.done(). This would result in the first PutObject to complete
completing the context. This has now been migrated to use async.each,
which will wait until all of the prefixes are processed before calling
a context completion event
* The code only supported us-west-2 - have made this use the Lambda
region value from the environment, and setup the interfaces similar to
what is done in the lambda loader
* Added references to configuration table constants rather than hard
coding