Version 2.7.9
Fixes a bug in how file sizes were reported in manifest creation, which caused Parquet based loads to fail. This version adds a new batch tracking data structure `entryMap` which contains not just the filename to be loaded, but also the size of the file obtained from the S3 event. These are then used to generate the Manifest `ContentLength` value. This version is backward compatible with 2.7.8 which structured batch entries as Lists of Strings, and will support hot upgrade.